反馈功能分离；修改record.log.logger.py中secure_view函数的raise_exc参数默认值，使得默认情况下根据DEBUG判断是否输出错误信息 (#760)
* 反馈功能分离；修改record.log.logger.py中secure_view函数的raise_exc参数默认值，使得默认情况下根据DEBUG判断是否输出错误信息

* 暂时取消通过反馈信息进行搜索的功能

---------
