+++ app/urls.py
@@ -16,7 +16,6 @@ from app import (
     org_views,
     activity_views,
     course_views,
-    feedback_views,
     academic_views,
     chat_api,
     YQPoint_views,
@@ -93,11 +92,6 @@ urlpatterns = [
     path("outputRecord/", course_views.outputRecord, name="outputRecord"),
     path("outputSelectInfo/", course_views.outputSelectInfo,
          name="outputSelectInfo"),
-] + [
-    # 反馈中心
-    path("feedback/", feedback_views.feedbackWelcome, name="feadbackWelcome"),
-    path("modifyFeedback/", feedback_views.modifyFeedback, name="modifyFeedback"),
-    path("viewFeedback/<str:fid>", feedback_views.viewFeedback, name="viewFeedback"),
 ] + [
     # 学术地图
     path("modifyAcademic/", academic_views.modifyAcademic, name="modifyAcademic"),
